SEC-0006: Refactor file structure
