+++ complete-node-bootcamp-master/4-natours/starter/controllers/userController.js
@@ -1,0 +1,37 @@
+const getAllUsers = (req, res) => {
+    res.status(500).json({
+        status: 'Route not implemented yet.'
+    });
+};
+
+const getUserById = (req, res) => {
+    res.status(500).json({
+        status: 'Route not implemented yet.'
+    });
+};
+
+const createUser = (req, res) => {
+    res.status(500).json({
+        status: 'Route not implemented yet.'
+    });
+};
+
+const updateUser = (req, res) => {
+    res.status(500).json({
+        status: 'Route not implemented yet.'
+    });
+};
+
+const deleteUser = (req, res) => {
+    res.status(500).json({
+        status: 'Route not implemented yet.'
+    });
+};
+
+module.exports = {
+    getAllUsers,
+    getUserById,
+    createUser,
+    updateUser,
+    deleteUser
+}
