+++ complete-node-bootcamp-master/4-natours/starter/package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start:dev": "node app.js",
-    "start:prod": "NODE_ENV=production nodemon server.js"
+    "start:dev": "node server.js",
+    "start:prod": "NODE_ENV=production node server.js"
   },
   "devDependencies": {
     "eslint": "^5.16.0",
