+++ complete-node-bootcamp-master/4-natours/starter/routes/tourRoutes.js
@@ -1,0 +1,15 @@
+const express = require('express');
+const tourController = require('./../controllers/tourController');
+
+const router = express.Router();
+
+router.route('/')
+    .get(tourController.getAllTours)
+    .post(tourController.createTour);
+
+router.route('/:id')
+    .get(tourController.getTourById)
+    .patch(tourController.updateTour)
+    .delete(tourController.deleteTour);
+
+module.exports = router;
