+++ complete-node-bootcamp-master/4-natours/starter/routes/userRoutes.js
@@ -1,0 +1,16 @@
+const express = require('express');
+
+const userController = require('./../controllers/userController');
+
+const router = express.Router();
+
+router.route('/')
+    .get(userController.getAllUsers)
+    .post(userController.createUser);
+
+router.route('/:id')
+    .get(userController.getUserById)
+    .patch(userController.updateUser)
+    .delete(userController.deleteUser);
+
+module.exports = router;
